+++ app/pages/search/_search-page.scss
@@ -13,4 +13,65 @@
     max-width: $app-SearchPage-max-width;
     z-index: $max-map-z-index +1 ;
   }
+
+  &__header {
+    margin-bottom: 20px;
+
+    .headroom {
+      top: 0;
+      zIndex: 1;
+    }
+    .headroom--unfixed {
+      position: relative;
+      transform: translateY(0);
+      margin-bottom: 20px;
+    }
+    .headroom--scrolled {
+      transition: transform 200ms ease-in-out;
+    }
+    .headroom--unpinned {
+      position: fixed;
+      transform: translateY(-100%);
+      max-width: $app-SearchPage-max-width;
+    }
+    .headroom--pinned {
+      position: fixed;
+      transform: translateY(0%);
+      background-color: rgba(0, 0, 0, 0.1);
+      z-index: 3000;
+      max-width: $app-SearchPage-max-width;
+      padding-top: 10px;
+    }
+
+    .app-MapToggle {
+      @include button-variant($gray, $white, $white);
+      display: block;
+      margin: 0 auto 20px auto;
+      box-shadow: 0px 1px 10px 1px rgba(0, 0, 0, 0.2);
+      border-radius: 100px;
+      padding-left: 30px;
+      padding-right: 40px;
+      font-size: 1.3rem;
+
+      & > div {
+        display: inline-block;
+      }
+
+      .app-MapToggle__icon {
+        color: $brand-primary;
+        font-size: 33px;
+        margin-right: 7px;
+      }
+
+      &__results-count {
+        font-size: 1.6rem;
+        color: $brand-primary;
+      }
+
+      &:hover, &:focus, &:active {
+        outline: none;
+        background-color: $white;
+      }
+    }
+  }
 }
