+++ app/pages/search/controls/_search-controls.scss
@@ -3,6 +3,8 @@
 }
 
 .app-SearchControlsContainer {
+  margin-bottom: 20px;
+
   .app-SearchBox {
     $button-width-sm: 60px;
     $button-width-lg: 100px;
