+++ app/pages/search/results/_search-results.scss
@@ -7,37 +7,6 @@
     margin-right: 0;
   }
 
-  .app-MapToggle {
-    @include button-variant($gray, $white, $white);
-    display: block;
-    margin: 0 auto;
-    box-shadow: 0px 1px 10px 1px rgba(0, 0, 0, 0.2);
-    border-radius: 100px;
-    padding-left: 30px;
-    padding-right: 40px;
-    font-size: 1.3rem;
-
-    & > div {
-      display: inline-block;
-    }
-
-    .app-MapToggle__icon {
-      color: $brand-primary;
-      font-size: 33px;
-      margin-right: 7px;
-    }
-
-    &__results-count {
-      font-size: 1.6rem;
-      color: $brand-primary;
-    }
-
-    &:hover, &:focus, &:active {
-      outline: none;
-      background-color: $white;
-    }
-  }
-
   .app-ResourceCompactList {
     position: fixed;
     bottom: 0;
